third pass in generate command
