+++ cmd/clean.go
@@ -1,8 +1,11 @@
 package cmd
 
 import (
+	"fmt"
+	"io/ioutil"
 	"os"
 	"path/filepath"
+	"strings"
 
 	"github.com/spf13/cobra"
 )
@@ -17,19 +20,48 @@ var cleanCmd = &cobra.Command{
 }
 
 func cleanAll() {
-	cwd, _ := os.Getwd()
+	rootDir, _ := os.Getwd()
+	deleteMigrationsDir(rootDir)
+	deleteRoutesGoFile(rootDir)
+	deleteMigrateDir(rootDir)
+	deleteTempDir(rootDir)
+	deleteAllGoFiles(rootDir)
+}
 
-	migration := filepath.Join(cwd, "migrations")
-	os.RemoveAll(migration)
-
-	routesGo := filepath.Join(cwd, "acl", "routes.go")
+func deleteRoutesGoFile(rootDir string) {
+	routesGo := filepath.Join(rootDir, "acl", "routes.go")
 	os.Remove(routesGo)
+}
 
-	bindataGo := filepath.Join(cwd, "migrate")
+func deleteMigrationsDir(rootDir string) {
+	migration := filepath.Join(rootDir, "migrations")
+	os.RemoveAll(migration)
+}
+
+func deleteMigrateDir(rootDir string) {
+	bindataGo := filepath.Join(rootDir, "migrate")
 	os.RemoveAll(bindataGo)
+}
 
-	tempDir := filepath.Join(cwd, "tmp")
+func deleteTempDir(rootDir string) {
+	tempDir := filepath.Join(rootDir, "tmp")
 	os.RemoveAll(tempDir)
+}
+
+func deleteAllGoFiles(rootDir string) error {
+	src := filepath.Join(rootDir, "schema")
+	files, err := ioutil.ReadDir(src)
+	if err != nil {
+		return err
+	}
+	for _, f := range files {
+		n := f.Name()
+		if strings.HasSuffix(n, ".go") {
+			fmt.Println(fmt.Sprintf("deleting %v", n))
+			os.Remove(filepath.Join(src, n))
+		}
+	}
+	return nil
 }
 
 func init() {
